+++ src/splunkParser.ts
@@ -4,8 +4,9 @@ import * as vscode from 'vscode';
 
 const commandInfoPath = "command_info";
 
-function returnCompletionItemfromJSON(fileName: string, filePath: string) {
-    let completeFilePath = path.join(filePath, `${fileName}+".csv"`);
+export function returnCompletionItemfromJSON(context: vscode.ExtensionContext, fileName: string) {
+    let completeFilePath = path.join(context.extensionPath, commandInfoPath, `${fileName}.json`);
+    
     console.log(`file path: "${completeFilePath}"`);
     let file: string;
     file = fs.readFileSync(completeFilePath, 'utf-8');
@@ -14,41 +15,45 @@ function returnCompletionItemfromJSON(fileName: string, filePath: string) {
     } catch(e) {
         console.log("Error: ", e.stack);
     }
-    console.log(file);
 
-    let fileJson = JSON.parse(file);
-    console.log(fileJson);
+    const fileJson = JSON.parse(file);
 
     // iterate through json entry and output completion items
     const name = "Command_Name";
-    const description = "Descrition";
+    const description = "Description";
     const syntax = "Supported functions and syntax";
     const related = "Related commands";
     const functionType = "Type of function";
+    const completionArray = [];
     
-    for (let entry of fileJson){
-        const commitCharacterCompletion = new vscode.CompletionItem('console');
-	    commitCharacterCompletion.commitCharacters = ['.'];
-	    commitCharacterCompletion.documentation = new vscode.MarkdownString('Press `.` to get `console.`');
+    for (const entry of fileJson) {
+        const splunkCompletionItem = new vscode.CompletionItem(entry[name]);
+        splunkCompletionItem.kind = vscode.CompletionItemKind.Keyword;
+        splunkCompletionItem.commitCharacters = ['\t'];
+
+        let hasSyntax: boolean = false;
+        let hasType: boolean = false;
+        let hasRelated: boolean = false;
+        if (functionType in entry) { hasType = true; }
+        if (syntax in entry) { hasSyntax = true; }
+        if (related in entry) { hasRelated = true; }
+
+        let detail: string = "";
+        let documentation: vscode.MarkdownString = new vscode.MarkdownString(entry[description]);
+        if ( hasRelated ) { documentation = new vscode.MarkdownString(documentation + "\n##### Related Commands:\n" + entry[related]); }
+        if ( hasType ) { detail = detail + '(' + entry[functionType] + ') '; }
+        if ( hasSyntax ) { detail = detail + entry[syntax]; }
+        
+        console.log(detail);
+        console.log(documentation);
+
+        splunkCompletionItem.detail = detail;
+        splunkCompletionItem.documentation = documentation;
+        completionArray.push(splunkCompletionItem);
     }
-    //return fileArray;
+    return completionArray;
 }
 
-function returnDictFromCSV(fileName: string, filePath: string) {
-    let completeFilePath = path.join(filePath, `${fileName}+".csv"`);
-    console.log(`file path: "${completeFilePath}"`);
-    let file;
-    try {
-        file = fs.readFileSync(completeFilePath, 'utf-8');
-    } catch(e) {
-        console.log("Error: ", e.stack);
-    }
-    console.log(file);
-
-    let fileArray = file?.split("\n");
-    console.log(fileArray);
-
-    return fileArray;
-}
-
-//export const mainFunctions = returnListFromCSV('Command', commandInfoPath);
+//export const mainFunctions = returnCompletionItemfromJSON('Command_description_Related_table', commandInfoPath);
+//export const evalFunctions = returnCompletionItemfromJSON('eval_functions-syntax_description_type', commandInfoPath);
+//export const statsFunctions = returnCompletionItemfromJSON('stats_functions-syntax_description_type', commandInfoPath);