+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "vscode-splunk-search-linter" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+## [1.0.0] - 2020-01-09
+### Added
+- Autocompleting commands
+- Syntax for eval and stats commands
+- Description for commands
+- Related main commands